Adding utils module for auxiliar functions.
