+++ react/tic-tac-toe/src/app/screens/Game/index.js
@@ -4,6 +4,8 @@ import styles from './styles.module.scss';
 
 import Board from './components/Board';
 
+import { calculateWinner, decideGameStatus } from '../../../utils/utils';
+
 class Game extends Component {
   constructor(props) {
     super(props);
@@ -72,37 +74,6 @@ class Game extends Component {
     </div>
   );
   }
-}
-
-function decideGameStatus(winner, xIsNext) {
-  let status;
-  if (winner) {
-    status = 'Winner: ' + winner;
-  } else {
-    status = 'Next player: ' + (xIsNext ? 'X' : 'O');
-  }
-  return status;
-}
-  
-
-function calculateWinner(squares) {
-  const lines = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6],
-  ];
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return squares[a];
-    }
-  }
-  return null;
 }
 
 export default Game;
